Fix: backend 파일 변환 후 삭제

+++ back/sleep/src/main/java/com/sleep/sleep/s3/S3Controller.java
@@ -7,8 +7,6 @@ import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.IOUtils;
-import org.springframework.core.io.InputStreamResource;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -19,8 +17,6 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.URLEncoder;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
@@ -44,16 +40,13 @@ public class S3Controller {
      * @return File 변환된 파일을 반환한다.
      * @throws IOException
      */
-    public static File multipartFileToFile(MultipartFile file) throws IOException {
-
-        File convFile = new File(file.getOriginalFilename());
-        convFile.createNewFile();
-        FileOutputStream fos = new FileOutputStream(convFile);
-        fos.write(file.getBytes());
-        fos.close();
-
+    public static File multipartFileToFile(MultipartFile multipartFile) throws IOException {
+        // 임시 디렉토리에 파일 생성
+        File convFile = File.createTempFile("upload-", "-" + multipartFile.getOriginalFilename());
+        try (FileOutputStream fos = new FileOutputStream(convFile)) {
+            fos.write(multipartFile.getBytes());
+        }
         return convFile;
-
     }
 
     @Operation(summary = "동영상 업로드", description = "추후 사용자별 업로드로 수정 예정; param: 업로드할 파일이름")
@@ -76,7 +69,12 @@ public class S3Controller {
         } catch (Exception e) {
             return new ResponseEntity<String>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
         } finally {
-            convFile.delete();
+            if (convFile != null && convFile.exists()) {
+                boolean isDeleted = convFile.delete();
+                if (!isDeleted) {
+                    System.err.println("Failed to delete the temporary file: " + convFile.getAbsolutePath());
+                }
+            }
         }
     }
 
@@ -144,15 +142,17 @@ public class S3Controller {
     }
 
     // 파일 로컬 저장
-    @PostMapping("/save/{fileName}")
-    public ResponseEntity<?> getLocalFilePath(@RequestHeader("Authorization") String accessToken, @PathVariable String fileName) {
+    @PostMapping("/save/{uploadNo}")
+    public ResponseEntity<?> getLocalFilePath(@RequestHeader("Authorization") String accessToken, @PathVariable int uploadNo) {
         try {
-            // 토큰에서 사용자 이름 추출 (이 코드는 사용자가 직접 구현해야 함)
             String username = jwtTokenUtil.getUsername(resolveToken(accessToken));
             //String encodedFileName = URLEncoder.encode(fileName, StandardCharsets.UTF_8.toString()).replaceAll("\\+", "%20");
+            System.out.println(username);
+
+            UploadShorts uploadShorts = uploadShortsRepository.findByUploadNo(uploadNo);
 
             // S3에서 파일 다운로드
-            InputStream inputStream = s3Service.downloadFile(username + "/" + fileName);
+            InputStream inputStream = s3Service.downloadFile(uploadShorts.getUploadTitle());
 
             // 임시 파일 생성
             File tempFile = File.createTempFile("downloaded-", ".mp4", new File(System.getProperty("java.io.tmpdir")));
@@ -161,9 +161,7 @@ public class S3Controller {
             // 파일로 스트림 복사
             Files.copy(inputStream, Paths.get(tempFile.getAbsolutePath()), StandardCopyOption.REPLACE_EXISTING);
 
-            // 작업 완료 로그
-            //System.out.println("File downloaded and saved to temporary directory: " + tempFile.getAbsolutePath());
-
+            // 블롭(바이트 배열) 반환
             return ResponseEntity.ok()
                     .body(tempFile.getAbsolutePath());
         } catch (Exception e) {
@@ -175,7 +173,6 @@ public class S3Controller {
     @PostMapping("/bring/blob/{fileName}")
     public ResponseEntity<?> getS3Blob(@RequestHeader("Authorization") String accessToken, @PathVariable String fileName) {
         try {
-            // 토큰에서 사용자 이름 추출 (이 코드는 사용자가 직접 구현해야 함)
             String username = jwtTokenUtil.getUsername(resolveToken(accessToken));
             //String encodedFileName = URLEncoder.encode(fileName, StandardCharsets.UTF_8.toString()).replaceAll("\\+", "%20");
 
@@ -185,7 +182,7 @@ public class S3Controller {
             // InputStream을 바이트 배열로 변환
             byte[] fileContent = IOUtils.toByteArray(inputStream);
 
-            // 블롭(바이트 배열)을 클라이언트에 반환
+            // 블롭(바이트 배열) 반환
             return ResponseEntity.ok()
                     .contentType(MediaType.APPLICATION_OCTET_STREAM)
                     .body(fileContent);
@@ -198,10 +195,10 @@ public class S3Controller {
     @PostMapping("/bring/myblob/{uploadNo}")
     public ResponseEntity<?> getUserS3Blob(@RequestHeader("Authorization") String accessToken, @PathVariable int uploadNo) {
         try {
-            // 토큰에서 사용자 이름 추출 (이 코드는 사용자가 직접 구현해야 함)
             String username = jwtTokenUtil.getUsername(resolveToken(accessToken));
             //String encodedFileName = URLEncoder.encode(fileName, StandardCharsets.UTF_8.toString()).replaceAll("\\+", "%20");
 
+            // 업로드 쇼츠 db에서 찾기
             UploadShorts uploadShorts = uploadShortsRepository.findByUploadNo(uploadNo);
 
             // S3에서 파일 다운로드
@@ -210,7 +207,7 @@ public class S3Controller {
             // InputStream을 바이트 배열로 변환
             byte[] fileContent = IOUtils.toByteArray(inputStream);
 
-            // 블롭(바이트 배열)을 클라이언트에 반환
+            // 블롭(바이트 배열) 반환
             return ResponseEntity.ok()
                     .contentType(MediaType.APPLICATION_OCTET_STREAM)
                     .body(fileContent);
